Rename ChatGLM1 to ChatGLM
