+++ intel_extension_for_transformers/llm/runtime/graph/application/main_run.cpp
@@ -296,7 +296,7 @@ int main(int argc, char** argv) {
     std::string prompt = build_prompt_glm2(prompts);
     embd_inp = ::model_tokenize(ctx, prompt, false);
     embd_inp.insert(embd_inp.begin(), {64790, 64792});  // special prefix
-  } else if (params.model_arch == MODEL_CHATGLM1) {
+  } else if (params.model_arch == MODEL_CHATGLM) {
     std::vector<std::string> prompts;
     prompts.push_back(params.prompt);
 
@@ -647,7 +647,7 @@ int main(int argc, char** argv) {
     }
 
     // display text
-    if (params.model_arch == MODEL_CHATGLM1 || params.model_arch == MODEL_CHATGLM2) {
+    if (params.model_arch == MODEL_CHATGLM || params.model_arch == MODEL_CHATGLM2) {
       static bool is_prompt = true;
       if (input_echo) {
         if (is_prompt == true) {
@@ -770,7 +770,7 @@ int main(int argc, char** argv) {
     }
 
     // end of text token
-    if (params.model_arch == MODEL_CHATGLM1) {
+    if (params.model_arch == MODEL_CHATGLM) {
       if (!embd.empty() && embd.back() == 130005) {
       if (params.instruct) {
           is_interacting = true;
